Refactor study_api, rename edit_participant URL
